requirements.txt
pydantic
kafka-python
